FAST_API/app/pyproject.toml
[tool.poetry]
name = "web_hw13_fast_api"
version = "0.1.0"
description = ""
authors = ["MicMaus <michael.yasynskyi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
SQLAlchemy = "^2.0.25"
psycopg = "^3.1.16"
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.5.3"}
bcrypt = "^4.1.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
PyJWT = "^2.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.13.1"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.1.0"
cloudinary = "^1.38.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
